requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
paho-mqtt==1.3.1
pymodbus==1.4.0
pyserial==3.4
requests==2.18.4
six==1.11.0
urllib3==1.22
